requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websocket-client>=1.6.0
pandas>=2.0.0
pyarrow>=10.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
aiofiles>=0.20.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0
numpy>=1.20.0
pytz>=2023.3